[frameworks] Rename frameworks to now-frameworks
