+++ packages/now-frameworks/logos/metalsmith.png
