+++ packages/now-frameworks/logos/pelican.png
